requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
cryptography
pillow
pytesseract
